+++ src/components/checkout/CheckoutForm.tsx
@@ -1,13 +1,16 @@
+"use client";
+
 import CurrencySelect from "@/components/checkout/CurrencySelect";
 import ReasonSelect from "@/components/checkout/ReasonSelect";
 import { CheckoutFormData, SinkCarbonXdrPostRequest } from "@/app/types";
 import Button from "@/components/Button";
 import { useAppContext } from "@/context/appContext";
-import { HTMLProps, useEffect, useState } from "react";
+import { HTMLProps, Suspense, useEffect, useState } from "react";
 import { SubmitHandler, useForm } from "react-hook-form";
 import CARBONCurrencyIcon from "@/components/icons/CARBONCurrencyIcon";
 import AmountInput from "@/components/checkout/AmountInput";
 import TransactionPreview from "@/components/checkout/TransactionPreview";
+import ParallaxDivider, { ParallaxBackgrounds } from "../ParallaxDivider";
 
 interface CheckoutFormProps extends HTMLProps<HTMLFormElement> {
   submitSinkingTransaction: (
@@ -29,10 +32,6 @@ export default function CheckoutForm({
   const currency = watch("currency");
   const reason = watch("reason");
 
-  useEffect(() => {
-    setValue("tonnes", 1);
-  }, [setValue]);
-
   const onSubmit: SubmitHandler<CheckoutFormData> = (data) => {
     let payload: SinkCarbonXdrPostRequest = {
       funder: walletConnection?.stellarPubKey!,
@@ -45,22 +44,29 @@ export default function CheckoutForm({
   };
 
   return (
-    <form className="flex flex-col gap-12 mx-5 md:mx-8 mt-10 bg-secondary min-w-[80%]">
-      <AmountInput
-        register={register}
-        watch={watch}
-        setValue={setValue}
-        quote={quote}
-        setQuote={setQuote}
-      />
-      <CurrencySelect register={register} />
-      <ReasonSelect watch={watch} setValue={setValue} />
-      <TransactionPreview
-        tonnes={tonnes}
-        currency={currency}
-        quote={quote}
-        onClick={() => handleSubmit(onSubmit)()}
-      />
+    <form className="">
+      <div className="flex flex-col gap-12 mx-5 md:mx-8 mt-10 bg-secondary min-w-[80%]">
+        <Suspense>
+          <AmountInput
+            register={register}
+            watch={watch}
+            setValue={setValue}
+            quote={quote}
+            setQuote={setQuote}
+          />
+        </Suspense>
+        <CurrencySelect register={register} />
+        <ReasonSelect watch={watch} setValue={setValue} />
+      </div>
+      <ParallaxDivider image={ParallaxBackgrounds.FOREST} smaller />
+      <div className="flex flex-col gap-12 mx-5 md:mx-8 mt-10 bg-secondary min-w-[80%]">
+        <TransactionPreview
+          tonnes={tonnes}
+          currency={currency}
+          quote={quote}
+          onClick={() => handleSubmit(onSubmit)()}
+        />
+      </div>
     </form>
   );
 }
